Updates pipeline for new environment.
The configure tempest function has been removed as now
configuration is done within the install tempest function as
we are able to install tempest outside of an LXC container
that uses mirror pip and apt repost.

Also removed an extra parameter on the run upgrade which was
previously used to call a function that returned failed upgrade
results that as used in testing the pipeline when the environment
was not able to be used.

Additionally, adds the .gitreview for this repo to use gerrit.

Change-Id: I2eb790f9cdf1614222d94fcac63f2c492628636b