Adds jUnit reports for test results
Adds converting last test results to jUnit style reports,
tranfsers the output files to the jenkins which then uses
the jUnit approach in the xUnit plugin to display trending
test results in a graph within the job.

Uses xUnit rather than jUnit for the ability to set the
unstable threshold, as using jUnit results in *any* skipped
or failed tests to cause the pipeline to report as unstable,
which is undesirable when some tests are expected to be skipped
or fail.

This makes it easier to review test results and patterns
from the tests results directly in the Jenkins pipeline.
Also eases confusion in how the tests are doing when the
pipeline displays as all green.

Change-Id: I3314ebcfa5203924bf571748fd580de8448b5bdd